backend/requirements.txt
# -------------------------
# Core API & typing
# -------------------------
fastapi
uvicorn
pydantic

# -------------------------
# Agentic orchestration
# -------------------------
langgraph

# -------------------------
# LLM providers (pluggable)
# -------------------------
openai
google-generativeai
requests

# -------------------------
# Environment management
# -------------------------
python-dotenv

prometheus-client

# -------------------------
# db
# -------------------------

sqlalchemy
